Split validations into a separate file
